apps/api/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Store
qdrant-client==1.7.3

# AI/ML
openai==1.10.0
# openai-agents-sdk==0.1.0  # Not available in PyPI - will implement manually if needed

# Authentication
# better-auth-python==0.1.0  # Not available in PyPI - will implement manually

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3
email-validator>=2.1.0
python-multipart==0.0.6

# Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Linting
flake8==7.0.0
